Created the videos app (#50)
* Moved the Videos model to videos app

* Added the videos app to installed appps and ran DB migrations

* Registered the Videos model to the videos app admin

* Updated URLconf for videos

* Updated views for videos

* Updated functional tests

* Updated docs

* Removed video tests data in resources model tests
