+++ docs/videos.rst
@@ -1,0 +1,53 @@
+videos package
+==============
+
+Submodules
+----------
+
+videos.admin module
+-------------------
+
+.. automodule:: videos.admin
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+videos.apps module
+------------------
+
+.. automodule:: videos.apps
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+videos.models module
+--------------------
+
+.. automodule:: videos.models
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+videos.urls module
+------------------
+
+.. automodule:: videos.urls
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+videos.views module
+-------------------
+
+.. automodule:: videos.views
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: videos
+   :members:
+   :undoc-members:
+   :show-inheritance:
