+++ resource_center/settings/base.py
@@ -54,6 +54,7 @@ LOCAL_APPS = [
     'resources',
     'accounts',
     'books',
+    'videos',
 ]
 
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS
